WASM building again, possible rework in std.json incoming
Former-commit-id: f6f166262850443fc0e95023810ccb6ea24db9ff
Former-commit-id: a437ad26477c0cc57ebf33f39fef50842499ac2a [formerly ea1c6c2391302e586a2d5497cb7d37ee3428abb2]
Former-commit-id: 79e78184a271039723a96fccb287f38d1fd0daad

+++ source/hip/graphics/g2d/spritebatch.d
@@ -32,7 +32,7 @@ public import hip.math.matrix;
     Vector2 tex_uv;
     int texID;
 
-    static enum floatCount = cast(ulong)(HipSpriteVertex.sizeof/float.sizeof);
+    static enum floatCount = cast(size_t)(HipSpriteVertex.sizeof/float.sizeof);
     static enum quadCount = floatCount*4;
     static assert(HipSpriteVertex.floatCount == 10,  "SpriteVertex should contain 9 floats and 1 int");
 }
